server/requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.35.0
pydantic==2.11.7
llama-cpp-python==0.3.9
faster-whisper==1.1.1
soundfile==0.13.1
numpy==2.3.2
